created working menu buttons for the home page and created reusable menu componennt that is utalized on the home page
